requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
numpy==1.26.4
PyYAML==6.0.1
scikit-learn==1.4.1.post1
scipy==1.12.0
threadpoolctl==3.3.0
Werkzeug==3.0.1
